+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Add functions emulator support for RTDB v2 triggers (#5045).
+- Enables single project mode for Firestore by default (#4890).